requirements.txt
requests
gitpython
argparse
giturlparse
